Factor out VirtualMembersList.
